% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enslide_plot.R
\name{enslide_list}
\alias{enslide_list}
\title{Generate a slide with a list of items}
\usage{
enslide_list(list, highlight, slide_caption, font_pct_size,
  font_family = "merriserif")
}
\arguments{
\item{font_pct_size}{}
}
\description{
Generate a slide with a list of items
}
