% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enslide_plot.R
\name{gk_slide}
\alias{gk_slide}
\title{Generate a General Knowledge slide}
\usage{
gk_slide(text, font_pct_size = 1)
}
\arguments{
\item{font_pct_size}{}
}
\value{

}
\description{
Generate a General Knowledge slide
}
