% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_hammer.R
\name{mc_hammer}
\alias{mc_hammer}
\title{Convert JSON quiz to HTML}
\usage{
mc_hammer(
  my_quiz,
  quiz_id,
  quiz_type = c("MultipleChoiceQuizz", "SingleChoiceQuizz")
)
}
\arguments{
\item{my_quiz}{Quiz as a string in JSON format}

\item{quiz_id}{Unique quiz id as a string}

\item{quiz_type}{MultipleChoiceQuizz or SingleChoiceQuizz, as character value}
}
\value{

}
\description{
Convert JSON quiz to HTML
}
