% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_hammer_list.R
\name{mc_hammer_list}
\alias{mc_hammer_list}
\title{Convert R list quiz to HTML}
\usage{
mc_hammer_list(my_quiz, quiz_id, quiz_type = c("MultipleChoiceQuizz",
  "SingleChoiceQuizz"))
}
\arguments{
\item{my_quiz}{Quiz as a list}

\item{quiz_id}{Unique quiz id as a string}

\item{quiz_type}{MultipleChoiceQuizz or SingleChoiceQuizz, as character value}
}
\description{
Convert R list quiz to HTML
}
