% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_hammer_test.R
\name{mc_hammer_test}
\alias{mc_hammer_test}
\title{Test parameters of the list are correct}
\usage{
mc_hammer_test(my_test_quiz)
}
\arguments{
\item{my_test_quiz}{Quiz as a list}
}
\value{

}
\description{
Test parameters of the list are correct
}
