% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_hammer_type_test.R
\name{mc_hammer_type_test}
\alias{mc_hammer_type_test}
\title{Test if single quiz or multiple quiz are done correctly}
\usage{
mc_hammer_type_test(my_quiz, quiz_type)
}
\arguments{
\item{my_quiz}{Quiz as a list}

\item{quiz_type}{MultipleChoiceQuizz or SingleChoiceQuizz, as character value}
}
\value{

}
\description{
Test if single quiz or multiple quiz are done correctly
}
