% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_lazy_quotes.R
\name{mc_lazy_quotes}
\alias{mc_lazy_quotes}
\title{Putting quotes for lazy people}
\usage{
mc_lazy_quotes(my_quiz)
}
\arguments{
\item{my_quiz}{Quiz as a string}
}
\description{
Putting quotes for lazy people
}
