% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_quiz.R
\name{preview_quiz}
\alias{preview_quiz}
\title{Include an image to preview the JSON quiz in markdown}
\usage{
preview_quiz(my_quiz)
}
\arguments{
\item{my_quiz}{Quiz as a string}
}
\description{
Include an image to preview the JSON quiz in markdown
}
