% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_quiz_list.R
\name{preview_quiz_list}
\alias{preview_quiz_list}
\title{Include an image to preview quiz}
\usage{
preview_quiz_list(my_quiz)
}
\arguments{
\item{my_quiz}{Quiz as list}
}
\description{
Include an image to preview quiz
}
