% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_fn.R
\name{star_fn}
\alias{star_fn}
\title{Title}
\usage{
star_fn(fn_name, pkg_name, top_caption = NULL,
  toolbox_type = c("shiny", "network", "timeseries", "cartography"),
  highlight = NULL, tweak_fn_size = 1)
}
\arguments{
\item{toolbox_type}{}
}
\description{
Title
}
