% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enslide_plot.R
\name{enslide_plot}
\alias{enslide_plot}
\title{Create a slide from a chart}
\usage{
enslide_plot(plot, slide_caption, plot_res = 160, plot_pct_width = 1,
  plot_align = "center")
}
\arguments{
\item{plot_align}{}
}
\description{
Create a slide from a chart
}
